New Sheriffs Office
Textured model with back. I did place it into the scene because I had to
scale it. For some reason the scaling process is never as consistent as
Alex let on. Its a different scale everytime I import the models even
though I'm working in a consistent scale in maya lol. Have a great day!